Moving components to subpackages
