+++ src/main/java/org/karnes/homebrew/emulator/component/bus/Bus.java
@@ -1,4 +1,4 @@
-package org.karnes.homebrew.emulator.component;
+package org.karnes.homebrew.emulator.component.bus;
 
 import org.karnes.homebrew.bitset.ArbitraryBitSet;
 import org.karnes.homebrew.bitset.FixedBitSet;
