+++ src/main/java/org/karnes/homebrew/emulator/component/bus/BusValueChangedEvent.java
@@ -1,4 +1,4 @@
-package org.karnes.homebrew.emulator.component;
+package org.karnes.homebrew.emulator.component.bus;
 
 import org.karnes.homebrew.bitset.FixedBitSet;
